♻️ refactors the stat area to use common components
